mythrilmerch-frontend/netlify/functions/api/requirements.txt
# mythrilmerch-frontend/netlify/functions/api/requirements.txt
Flask
Flask-Cors
psycopg2-binary